chore: Form inputs refactor
