+++ packages/web/src/formInputs/textFields/numberField.ts
@@ -1,19 +1,39 @@
 import { NumberFieldElement } from '@vaadin/vaadin-text-field/src/vaadin-number-field'
 import { mix } from 'mixwith'
-import { EntityStoreMixin, EntityStorePathMixin } from '../mixins'
+import { EntityStoreMixin, EntityStorePathMixin } from '../../mixins'
+
+/**
+ * `<c6o-number-field>` is a Web Component for number field control in forms.
+ *
+ * ```html
+ * <c6o-number-field label="Number"></c6o-number-field>
+ *
+ * <c6o-number-field currency label="Cost"></c6o-number-field>
+ * ```
+ *
+ * The following state attributes are available for styling:
+ *
+ * Attribute    | Description | Part name
+ * -------------|-------------|------------
+ * `currency  ` | Set when the number input is a currency to 2 decimal places | :host
+ *
+ * @extends EntityStorePathMixin
+ * @mixes NumberFieldElement
+ * @mixes EntityStoreMixin
+ * @mixes EntityStorePathMixin
+ */
 
 export interface NumberField extends EntityStorePathMixin {
+    value: string
 }
 
 export class NumberField extends mix(NumberFieldElement).with(EntityStoreMixin, EntityStorePathMixin) {
     currency: boolean
-    value: string
 
     static get properties() {
         return {
             ...super.properties,
             currency: { type: Boolean },
-            value: { type: String }
         }
     }
