+++ packages/web/src/grid/grid.ts
@@ -169,6 +169,7 @@ import { mix } from 'mixwith'
 export interface Grid extends EntityListStoreMixin {
     activeItem: boolean
     getEventContext(e)
+    items: Array<any>
 }
 
 export class Grid extends mix(GridElement).with(EntityListStoreMixin) {
@@ -185,6 +186,9 @@ export class Grid extends mix(GridElement).with(EntityListStoreMixin) {
 
     async connectedCallback() {
         await super.connectedCallback()
+
+        if (this.items?.length && this.listStore)
+            throw new Error('Both an items array and a ListStore were provided to the Grid component. Please only use one or the other.')
 
         this.$.table.addEventListener('scroll', this.handleScroll)
