+++ packages/web/src/index.ts
@@ -12,27 +12,15 @@ export * from './badge'
 export * from './banners/'
 export * from './buttons/'
 export * from './details'
-export * from './dialogs/baseDialog'
-export * from './dialogs/baseStoreDialog'
-export * from './dialogs/confirmDialog'
-export * from './dialogs/dialog'
-export * from './dialogs/errorDialog'
-export * from './dialogs/infoDialog'
-export * from './dialogs/storeDialog'
-export * from './form/layout'
-export * from './form/item'
-export * from './form/results'
-export * from './formInputs/checkbox'
+export * from './dialogs/'
+export * from './form/'
+export * from './formInputs/checkbox/'
 export * from './formInputs/combobox'
-export * from './formInputs/numberField'
-export * from './formInputs/passwordField'
-export * from './formInputs/radioButton'
-export * from './formInputs/radioGroup'
-export * from './formInputs/select'
-export * from './formInputs/textArea'
-export * from './formInputs/textField'
+export * from './formInputs/radios/'
+export * from './formInputs/select/'
+export * from './formInputs/textFields/'
 export * from './formInputs/toggle'
-export * from './formInputs/upload'
+export * from './formInputs/upload/'
 export * from './grid/'
 export * from './label'
 export * from './layout/accordion'
